display of text / comment / boolean / color / dropdown / radiogroup working #11181
