+++ projects/safe/src/lib/components/ui/grid-core/grid/grid.component.ts
@@ -1,5 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { GridDataResult } from '@progress/kendo-angular-grid';
+import { SafeExpandedCommentComponent } from '../expanded-comment/expanded-comment.component';
+import get from 'lodash/get';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'safe-grid',
@@ -13,9 +16,28 @@ export class SafeGridComponent implements OnInit {
   @Input() data: GridDataResult = { data: [], total: 0 };
   @Input() loading = false;
 
-  constructor() { }
+  constructor(private dialog: MatDialog) { }
 
   ngOnInit(): void {
+  }
+
+  // === DATA ===
+  /**
+   * Displays text instead of values for questions with select.
+   * @param meta meta data of the question.
+   * @param value question value.
+   * @returns text value of the question.
+   */
+  public getDisplayText(value: string | string[], meta: { choices?: { value: string, text: string }[] }): string | string[] {
+    if (meta.choices) {
+      if (Array.isArray(value)) {
+        return meta.choices.reduce((acc: string[], x) => value.includes(x.value) ? acc.concat([x.text]) : acc, []);
+      } else {
+        return meta.choices.find(x => x.value === value)?.text || '';
+      }
+    } else {
+      return value;
+    }
   }
 
   // === UTILITIES ===
@@ -24,7 +46,7 @@ export class SafeGridComponent implements OnInit {
    * @param e Component resizing event.
    * @returns True if overflows.
    */
-   isEllipsisActive(e: any): boolean {
+  isEllipsisActive(e: any): boolean {
     return (e.offsetWidth < e.scrollWidth);
   }
 
@@ -34,6 +56,30 @@ export class SafeGridComponent implements OnInit {
    * @param rowTitle field name.
    */
   public onExpandText(item: any, field: any): void {
-    console.log('on expand');
+    const dialogRef = this.dialog.open(SafeExpandedCommentComponent, {
+      data: {
+        title: field.title,
+        comment: get(item, field.name)
+      },
+      autoFocus: false,
+      position: {
+        bottom: '0',
+        right: '0'
+      },
+      panelClass: 'expanded-widget-dialog'
+    });
+    dialogRef.afterClosed().subscribe(res => {
+      // TODO: finish that
+      // if (res !== item[rowTitle]) {
+      //   this.gridData.data.find(x => x.id === item.id)[rowTitle] = res;
+      //   this.items.find(x => x.id === item.id)[rowTitle] = res;
+      //   if (this.updatedItems.find(x => x.id === item.id) !== undefined) {
+      //     this.updatedItems.find(x => x.id === item.id)[rowTitle] = res;
+      //   }
+      //   else {
+      //     this.updatedItems.push({ [rowTitle]: res, id: item.id });
+      //   }
+      // }
+    });
   }
 }
